Reuse decompression contexts
This change allows one ntfs_system_decompression_ctx to be reused for
all reads from a given open FUSE file.  To make this possible, it was
necessary to update the decompression context to no longer hold a
reference to the open compressed stream.

Based on a proposal by Mikalai Ramanovich.