Avoid -Waddress-of-packed-member warning
Add __attribute__((packed)) to the on-disk reparse point structures to
avoid a compiler warning about a potential unaligned pointer when
casting a pointer to REPARSE_POINT (which is defined in
<ntfs-3g/layout.h> using __attribute__((packed))) to a pointer to
WOF_FILE_PROVIDER_REPARSE_POINT_V1.

This warning is probably a false positive since the REPARSE_POINT from
NTFS-3G should be aligned in practice.  But we shouldn't assume that.